feat(io): Add restart functionality via checkpoints and switch to time-based saving
Initial implementation of saving and loading capability by specifying a bool value in the config file.
Also switched the saving from step-based to time-based, specifying two save_dt in the config file, one
for the initial steps and one for the refined steps with a trigger time for the switch.